requirements.txt
pytest
coverage
Flask
flask-cors
Flask==2.3.3
gunicorn==21.2.0
python-dotenv

